[step3] refactor: 훅에서 비즈니스 로직을 분리하고 종속적인 훅으로 이동
