+++ docs/REQUIREMENTS.md
@@ -128,4 +128,4 @@ step3
 
 - [ ] 모달을 분리하여 npm으로 배포하고, 그 라이브러리를 직접 import해서 사용
 - [ ] '카드를 등록중입니다' 스피너 추가
-- [ ] custom Hook 비즈니스 로직 분리
+- [x] custom Hook 비즈니스 로직 분리
