+++ src/pages/CardNamePage/index.tsx
@@ -1,37 +1,11 @@
-import CardPreview from 'pages/RegisterPage/CardPreview';
 import { StyledButton } from 'components/Button';
 import { StyledInput } from 'components/Input';
-import { AddCardContext } from 'context/CardContext';
-import { useContext } from 'react';
-import { useNavigate } from 'react-router-dom';
+import { useCardNameHandler } from './hooks/useCardNameHandler';
+import CardPreview from 'pages/RegisterPage/CardPreview';
 import styled from 'styled-components';
-import { setLocalStorage } from 'utils/localStorage';
-import { Card } from 'types/Card';
 
 const CardNamePage = () => {
-  const navigate = useNavigate();
-
-  const { cardNumber, date, name, cardCompany, cardName, setCardName } =
-    useContext(AddCardContext);
-
-  const cardInfo: Card = {
-    cardNumber: cardNumber,
-    date: date,
-    name: name,
-    cardCompany: cardCompany,
-    cardName: cardName,
-  };
-
-  const handleCardName = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const { value } = event.target;
-
-    setCardName?.(value);
-  };
-
-  const onSubmitHandler = () => {
-    setLocalStorage('card', { ...cardInfo });
-    navigate('/');
-  };
+  const { cardInfo, handleCardName, onSubmitHandler } = useCardNameHandler();
 
   return (
     <>
@@ -41,7 +15,7 @@ const CardNamePage = () => {
         <form onSubmit={onSubmitHandler}>
           <CardNameInput
             autoFocus
-            value={cardName}
+            value={cardInfo.cardName}
             onChange={handleCardName}
             required
           ></CardNameInput>
